Deploy ADE to Virtual Machines (#144)
Lots of changes to accommodate the VMs properly:

- Added App Configuration Service
- Updated Virtual Machines with Custom Script Extensions to install containers and run them
- Updated App to Use App Config for main configuration
- Updated App Services to also use App Config
- Added multiple app gateway endpoints for each of the compute scenarios
- Some cleanup around execution order
- Added Github Actions to push to Docker Hub

Known Issue: when accessing the VM app gateway endpoint, a CORS error comes up.
Known Issue: occasionally, an az cli error comes up saying that "byte is not a property of object". Not sure if this is from us, or from az.